feat(*): add feed item status
