+++ src/components/feed-list-item/feed-list-item.tsx
@@ -1,20 +1,19 @@
+import { FeedListItemStatus } from '../feed-list/types';
 import { Price } from '../price';
+import { FeedListItemProps } from './types';
 import styles from './styles.module.css';
 
-export type FeedListItemProps = {
-    item: {
-        name: string;
-        time: string;
-        number: string;
-        ingredients: {
-            bunId: string;
-            ingredientsIds: Array<string>;
-        };
-        price: number;
-    };
-};
-
 export const FeedListItem = ({ item }: FeedListItemProps) => {
+    const statusClass: string = getStatusClass(item.status);
+
+    function getStatusClass(status?: FeedListItemStatus): string {
+        if (status === FeedListItemStatus.COMPLETED) {
+            return `text_type_main-default ${styles['item-status-completed']}`;
+        }
+
+        return `text_type_main-default ${styles['item-status']}`;
+    }
+
     return (
         <div className={styles['item-container']}>
             <div className={styles['item-title']}>
@@ -25,9 +24,14 @@ export const FeedListItem = ({ item }: FeedListItemProps) => {
                     {item.time}
                 </div>
             </div>
-            <div className='text text_type_main-medium'>{item.name}</div>
-            <div>
-                <div>
+            <div className='text text_type_main-medium'>
+                {item.name}
+                {item.status && (
+                    <div className={statusClass}>{item.status}</div>
+                )}
+            </div>
+            <div className={styles['item-content-and-price']}>
+                <div className={styles['item-content']}>
                     {item.ingredients.bunId}
                     {item.ingredients.ingredientsIds}
                 </div>
