+++ src/components/feed-list-item/types.ts
@@ -1,0 +1,5 @@
+import { FeedListItemContent } from "../feed-list/types";
+
+export type FeedListItemProps = {
+    item: FeedListItemContent;
+};
